requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-auth==2.35.0
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.3
pyarrow==17.0.0
db-dtypes==1.3.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
ruff==0.6.9
